Make JupyterLab page actually work in browser window
We have to turn of node integration to keep from interferring with
JupyterLabs' JS.  JupyterLab uses a beforeunload handler to check if you
really want to leave the page.  Electron doesn't support that correctly
(https://github.com/electron/electron/issues/2579), so we catch close
events, pop up our own dialog, and destroy the window if an exit is
desired.

Also, load the full URL specified, as that now has an authentication
token.